rename NamespaceReplacerRector to RenameNamespaceRector
