+++ tests/Rector/Namespace_/RenameNamespaceRector/Fixture/fixture.php.inc
