+++ tests/Rector/Namespace_/RenameNamespaceRector/Fixture/fixture2.php.inc
