+++ tests/Rector/Namespace_/RenameNamespaceRector/Fixture/fixture3.php.inc
