+++ tests/Rector/Namespace_/RenameNamespaceRector/Fixture/fixture4.php.inc
