+++ tests/Rector/Namespace_/RenameNamespaceRector/Fixture/fixture5.php.inc
